move: extract bulk of logic to function for reuse by `jj squash`
